py3 compatible + setup.py
