+++ surfvis/__init__.py
@@ -1,0 +1,8 @@
+"""
+Per-baseline time freq plots
+
+author - Ian Heywood
+email  - ianh@astro.ox.ac.uk
+date   - 02/08/2021
+"""
+__version__ = '0.0.1'
